expose all preparation things for run with js
