+++ aqua-run/src/main/scala/aqua/run/CallInfo.scala
@@ -1,0 +1,9 @@
+package aqua.run
+import aqua.definitions.FunctionDef
+
+case class CallInfo(
+  name: String,
+  air: String,
+  definitions: FunctionDef,
+  config: RunConfig
+)
